Move keywords to syntax
