+++ berg-compiler/src/eval/mod.rs
@@ -4,7 +4,6 @@ mod root;
 
 pub use self::block::BlockRef;
 pub use self::expression_eval::ExpressionEvaluator;
-pub use self::root::keywords;
 pub use self::root::RootRef;
 
 use crate::syntax::AstRef;
