+++ berg-compiler/src/syntax/identifiers.rs
@@ -146,3 +146,14 @@ impl fmt::Debug for IdentifierIndex {
         }
     }
 }
+
+///
+/// Keywords are fields in the root. When the identifier `true` is in the code,
+/// it's treated as a normal variable reference and looked up in scope (which
+/// includes the root scope).
+///
+#[allow(clippy::upper_case_acronyms)]
+pub mod keywords {
+    use crate::fields;
+    fields! { TRUE, FALSE, IF, ELSE, WHILE, FOREACH, BREAK, CONTINUE, TRY, CATCH, FINALLY, THROW, }
+}
